Fix ConcurrencyViolation on save with some spells
The UPDATE command was checking every column to see if any data had changed since the SELECT statement had been performed. Some data has changed since I default NULL data to a standard value, which the WoW client reads the same way. To fix this, I set the conflict resolution method to overwrite - so it only checks the primary key. The downside of this is that it will overwrite other peoples changes if they saved changes after you loaded it, but I think this is acceptable for now.