Fix export to DBC & unhandled exception:
The export to DBC functionality was broken in the same way as the previous commit. A connection already closed exception is raised when the program is terminated gracefully - I have not investigated why this is the case, maybe because there is no external connection, so I just let it exit gracefully in this case.