Implement connection type configuration, fix startup error
You can now choose between initating a MySQL or SQLite connection when starting the spell editor. SQLite is a local database that doesn't require anything to use. You can share this database created with others and export it back to a DBC file.

The MySQL version connects to a remote server such that multiple people can be editing the same database. It also then provides the ability to fully query the database and access it with GUI programs.

A startup error was happening because of some debug code left in and has now been resolved.