Make the DBC and icon loading process display in the UI
This is to make it more obvious the program is doing something and has not just crashed.

A better solution would be to refactor the code so that the DBC and icon loading does not directly update the UI, the UI can be updated on the main thread after everything required has been loaded into memory asynchronous.