When selecting spells from the database order by the primary key
When creating a new spell with a low spell id I was annoyed at the spell appearing at the end of the list rather than in the ordered position. In the database it will still be the last row but when you export back to a DBC file it reorders it.