Delete the BlpReference class that is no longer used
Currently images that are no longer used (because they are not visible in the UI) are never garbage collected because the reference persists in the BlpManager ImageMap. To be more effecient here we could count the number of current references visible, and if that reaches zero then delete the source from the map.