Move Icon tab image loading to another thread for performance gain
The images will be asynchronously loaded and updated in the UI so that the UI doesn't freeze while the images are loading.

This is needed because IsVisibleChanged fires immediately for every image in the WrapPanel. I have been unable to find a way to detect images coming into view as you scroll using this control.