Fix some UI elements not updating after changing language
This implementation clears UI elements that contain multiple items and rebuilds them in order to pick up the new language definitions. This is ineffecient because we are creating more UI elements each change. A better solution would be to reuse existing elements and just change the display properties. However this implementation takes longer to implement as it involves a rewrite of this section.