Fix exception parsing strings like ${$100d / 6} and improve error logging
The spell string ${$100d / 6} was raising an exception because $100d would resolve to "10 seconds" from spell ID 100 and then that value could not be converted to a double because of the seconds string. To work around this I have hacked it by returning the first portion of the string only, splitting by space character, in order to retrieve the numerical value.

The error logging has been improved to include the stack trace in the error log and to include the exception type in the UI message.