Make SpellDifficulty tooltips be loaded on background thread
This is a performance improvement when connecting to a remote MySQL server. It would freeze the UI for multiple minutes while all of the spell difficulty tooltips were built.

The algorithm used is still quite ineffecient, but it is now on a background thread and loaded asynchronously. This stops the UI hanging.