Fix #155 SpellVisualKitModelAttach with AttachmentId -1
When a spell visual kit model attachment with an AttachmentId of -1 was selected it would throw a hard erorr in the UI and fail to load the attachment.

This fixes the error, but does not fix the root cause of the issue. The database schema currently has the wrong data type, it stores it as a uint rather than a int.