#216 #220 Fix spell list not refreshing correctly
This occurred when saving a spell becausue it would update it in the UI but not in the in-memory data table. So when the entry was next refreshed, it would refresh with the wrong information.

This change refactors the behaviour into the appropiate class and handles the data and UI update properly.