Refactor startup logic and dependencies (#235)
This reduces the time it takes for the spell editor to startup significantly.

To achieve this, the logic being initialised can be processed concurrently. All UI dependent code is isolated which must go through the UI thread.

The SQL adapters use locking when executing queries which slows down the startup. To avoid this temporary adapters are created for the use of the startup logic.